// src/api/generated/model/sportPreferenceResponseDataSport.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Ludora API
 * API for the Ludora app
 * OpenAPI spec version: 0.0.1
 */

/**
 * The sport of the user sport preference
 */
export type SportPreferenceResponseDataSport =
  (typeof SportPreferenceResponseDataSport)[keyof typeof SportPreferenceResponseDataSport];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SportPreferenceResponseDataSport = {
  BASKETBALL: 'BASKETBALL',
  FOOTBALL: 'FOOTBALL',
  TENNIS: 'TENNIS',
  VOLLEYBALL: 'VOLLEYBALL',
  PADDEL: 'PADDEL',
  BADMINTON: 'BADMINTON',
  'PING-PONG': 'PING-PONG',
} as const;